+++ flow/connectors/clickhouse/type_conversion.go
@@ -2,40 +2,13 @@ package connclickhouse
 
 import (
 	"github.com/PeerDB-io/peerdb/flow/generated/protos"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
+	"github.com/PeerDB-io/peerdb/flow/shared/clickhouse"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 )
-
-/*
-This file handles the mapping for ClickHouse destination types and
-their corresponding TypeConversion implementations. A TypeConversion
-object contains two functions: one for schema conversion (QField) and
-one for value conversion (QValue). This allows the avro writer to
-stage the schema/data in the converted type format, and therefore
-successfully uploaded to the desired destination type in ClickHouse.
-
-To add a type conversion:
-	(1) In flow/model/qvalue/type_converter.go:
-	- implement a SchemaConversionFn interface to convert the QField type
-	- implement a ValueConversionFn interface to convert the QValue data
-
-	(2) Add the new conversion to the `supportedDestinationTypes` map here
-		(if destination type doesn't exist, create a new map entry for it).
-
-The GetColumnsTypeConversion function returns the full list of supported
-type conversions. Note that the source types are QValueKind, this allows
-the implementation to be source-connector agnostic.
-*/
-
-var supportedDestinationTypes = map[string][]qvalue.TypeConversion{
-	"String": {qvalue.NewTypeConversion(
-		qvalue.NumericToStringSchemaConversion,
-		qvalue.NumericToStringValueConversion,
-	)},
-}
 
 func GetColumnsTypeConversion() (*protos.ColumnsTypeConversionResponse, error) {
 	res := make([]*protos.ColumnsTypeConversion, 0)
-	for qkind, destTypes := range listSupportedTypeConversions() {
+	for qkind, destTypes := range clickhouse.ListSupportedTypeConversions() {
 		res = append(res, &protos.ColumnsTypeConversion{
 			Qkind:            string(qkind),
 			DestinationTypes: destTypes,
@@ -44,17 +17,6 @@ func GetColumnsTypeConversion() (*protos.ColumnsTypeConversionResponse, error) {
 	return &protos.ColumnsTypeConversionResponse{
 		Conversions: res,
 	}, nil
-}
-
-func listSupportedTypeConversions() map[qvalue.QValueKind][]string {
-	typeConversions := make(map[qvalue.QValueKind][]string)
-
-	for dstType, l := range supportedDestinationTypes {
-		for _, conversion := range l {
-			typeConversions[conversion.FromKind()] = append(typeConversions[conversion.FromKind()], dstType)
-		}
-	}
-	return typeConversions
 }
 
 func findTypeConversions(schema qvalue.QRecordSchema, columns []*protos.ColumnSetting) map[string]qvalue.TypeConversion {
@@ -70,7 +32,7 @@ func findTypeConversions(schema qvalue.QRecordSchema, columns []*protos.ColumnSe
 		if !exist {
 			continue
 		}
-		conversions, exist := supportedDestinationTypes[col.DestinationType]
+		conversions, exist := clickhouse.SupportedDestinationTypes[col.DestinationType]
 		if !exist {
 			continue
 		}
