+++ flow/connectors/mysql/qvalue_convert.go
@@ -17,8 +17,8 @@ import (
 
 	"github.com/PeerDB-io/peerdb/flow/datatypes"
 	"github.com/PeerDB-io/peerdb/flow/generated/protos"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
 	"github.com/PeerDB-io/peerdb/flow/shared"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 )
 
 func qkindFromMysql(field *mysql.Field) (qvalue.QValueKind, error) {
