+++ flow/connectors/postgres/qvalue_convert.go
@@ -16,8 +16,8 @@ import (
 	"github.com/shopspring/decimal"
 
 	datatypes "github.com/PeerDB-io/peerdb/flow/datatypes"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
 	"github.com/PeerDB-io/peerdb/flow/shared"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 )
 
 func (c *PostgresConnector) postgresOIDToName(recvOID uint32, customTypeMapping map[uint32]shared.CustomDataType) (string, error) {
