+++ flow/connectors/postgres/schema_delta_test_constants.go
@@ -2,7 +2,7 @@ package connpostgres
 
 import (
 	"github.com/PeerDB-io/peerdb/flow/generated/protos"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 )
 
 var AddAllColumnTypesFields = []*protos.FieldDescription{
