+++ flow/connectors/snowflake/avro_file_writer_test.go
@@ -14,7 +14,7 @@ import (
 	avro "github.com/PeerDB-io/peerdb/flow/connectors/utils/avro"
 	"github.com/PeerDB-io/peerdb/flow/generated/protos"
 	"github.com/PeerDB-io/peerdb/flow/model"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 )
 
 // createQValue creates a QValue of the appropriate kind for a given placeholder.
