+++ flow/e2e/bigquery/peer_flow_bq_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/PeerDB-io/peerdb/flow/e2e"
 	"github.com/PeerDB-io/peerdb/flow/e2eshared"
 	"github.com/PeerDB-io/peerdb/flow/generated/protos"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
 	"github.com/PeerDB-io/peerdb/flow/shared"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 	peerflow "github.com/PeerDB-io/peerdb/flow/workflows"
 )
