+++ flow/model/record_items.go
@@ -6,7 +6,7 @@ import (
 	"math"
 
 	"github.com/PeerDB-io/peerdb/flow/datatypes"
-	"github.com/PeerDB-io/peerdb/flow/model/qvalue"
+	"github.com/PeerDB-io/peerdb/flow/shared/qvalue"
 )
 
 type Items interface {
